add header to page templates
adding header to page templates to provide context for users on the provenance of this content. text is as follows:

This content was recently an official website of the City of Austin. It was available at odd.austintexas.io, until the CNAME file was deleted on July 20th, 2021, as seen in this Github commit history.

In the spirit of transparency and open source, we’ve re-published its content here. You can also view it with the former domain using web.archive.org